PROG: Minor fixes to warnings/compile errors on various other OS's caused by #1312. References #1312.
git-svn-id: http://subversion.wr.usgs.gov/repos/prog/isis3/trunk@4930 41f8697f-d340-4b68-9986-7bafba869bb8

+++ isis/src/base/apps/cubediff/cubediff.cpp
@@ -279,7 +279,7 @@ void diffTable(ofstream &target, int precision) {
     Column currCol;
     //Prepare and add the first file's column
     currCol.SetName(QString("File1_") + toString(samps[i]));
-    if((unsigned int)(colWidth + precision + 1) < currCol.Name().length()) {
+    if((int)(colWidth + precision + 1) < (int)currCol.Name().length()) {
       currCol.SetWidth(currCol.Name().length() + 1);
     }
     else currCol.SetWidth(colWidth + precision + 1);
